+++ src/schema/query/process.js
@@ -1,0 +1,13 @@
+/**
+ * @memberof Eventsource.Schema
+ * @typedef {Object} QueryProcess
+ */
+const schema =
+{
+  '@meta': 
+  {
+    'extends' : 'eventsource/schema/entity/process'
+  }
+}
+
+module.exports = schema